fixed Emoji creation after Coingecko recalc
Was in the wrong place =) (Needs to be rechecked and overwritten for each new datasource) 
If there are more than 2 Datasources in the future the logic how to average might need to be rewritten. 

(In the longrun maybe build each 2D array the same way to itterate over all availlable ones? Maybe with an "indexarray" that keeps track of all choosen Datasources to then iteate over them)